Fix to determine_move function
I did two things:
First, made sure instead of invalid move, determine_move function just returns wait.
Second, I added a boolean parameter to determine_move (default to False) which, if True, with 50% possibility, lets the player move towards the target in an open diagonal direction in case the first cell of the shortest path is blocked. This may potentially fasten the search, but I set it to Default:False as I did not extensively test it yet.